backend/apps/patients/serializers.py
from rest_framework import serializers
from .models import Patient
from apps.users.serializers import UserSerializer

class PatientSerializer(serializers.ModelSerializer):
    user = UserSerializer(read_only=True)

    class Meta:
        model = Patient
        fields = ['id', 'user', 'date_of_birth', 'blood_group', 'address']